+++ system/src/templates.mjs
@@ -138,7 +138,6 @@ export default function() {
 		"systems/shadowdark/templates/items/language/details-tab.hbs",
 		"systems/shadowdark/templates/items/npc-attack/details-tab.hbs",
 		"systems/shadowdark/templates/items/npc-special-attack/details-tab.hbs",
-		"systems/shadowdark/templates/items/npc-spell/details-tab.hbs",
 		"systems/shadowdark/templates/items/patron/details-tab.hbs",
 		"systems/shadowdark/templates/items/potion/details-tab.hbs",
 		"systems/shadowdark/templates/items/property/details-tab.hbs",